refactor: refactorred policy for all controllers, models, resources

+++ app/Http/Controllers/Backend/MediaController.php
@@ -21,7 +21,7 @@ class MediaController extends Controller
 
     public function index(Request $request)
     {
-        $this->checkAuthorization(Auth::user(), ['media.view']);
+        $this->authorize('viewAny', Media::class);
 
         // Check for PHP upload limit errors first
         $phpError = MediaHelper::checkPhpUploadError();
@@ -84,7 +84,7 @@ class MediaController extends Controller
 
     public function store(MediaUploadRequest $request)
     {
-        $this->checkAuthorization(Auth::user(), ['media.create']);
+        $this->authorize('create', Media::class);
 
         if (config('app.demo_mode', false) && Media::count() > 10) {
             return response()->json([
@@ -132,7 +132,8 @@ class MediaController extends Controller
 
     public function destroy($id)
     {
-        $this->checkAuthorization(Auth::user(), ['media.delete']);
+        $media = Media::findOrFail($id);
+        $this->authorize('delete', $media);
 
         $this->mediaLibraryService->deleteMedia($id);
 
@@ -144,7 +145,7 @@ class MediaController extends Controller
 
     public function bulkDelete(MediaBulkDeleteRequest $request)
     {
-        $this->checkAuthorization(Auth::user(), ['media.delete']);
+        $this->authorize('bulkDelete', Media::class);
 
         $this->mediaLibraryService->bulkDeleteMedia($request->ids);
 
@@ -153,7 +154,7 @@ class MediaController extends Controller
 
     public function api(Request $request)
     {
-        $this->checkAuthorization(Auth::user(), ['media.view']);
+        $this->authorize('viewAny', Media::class);
 
         $result = $this->mediaLibraryService->getMediaList(
             $request->get('search'),
